refactor(storage): address review feedback on move command
- Merge storage_move_utils.go into storage_move.go; no other storage
  command has a _utils.go split.
- Inline the confirmPrefixMove wrapper; the call is a single AskQuestion
  expression.
- Handle flag-lookup errors in RunE instead of discarding them with _,
  matching the pattern used by every other storage command.
- Replace return cmd.Usage() in PreRunE with
  exocmd.CmdExitOnUsageError, matching storage_delete and other commands
  that exit with a 'invalid arguments' reason on bad arity.
